working version-1.0 Features: -Map added -Displaying the inCycle truck operators on the map. -Delete cycle option added
Enhancements:
-Realtime adding of the new truck operators to the map
-Realtime position change of the truck operators on the map
-Displaying the truck operators only after receiving the initial coordinate value.
