+++ app/src/main/java/com/example/plantoperator/Adapters/ProgrammingAdapter.java
@@ -1,4 +1,4 @@
-package com.example.plantoperator;
+package com.example.plantoperator.Adapters;
 
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -10,6 +10,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.plantoperator.R;
+
 import java.util.List;
 
 public class ProgrammingAdapter extends RecyclerView.Adapter<ProgrammingAdapter.ProgrammingViewHolder> {
@@ -20,6 +22,7 @@ public class ProgrammingAdapter extends RecyclerView.Adapter<ProgrammingAdapter.
         Log.d("recycler", "insideConstructorPA: " + this.data);
         this.listener = listener;
     }
+
 
     @NonNull
     @Override
